- **boss.cpp / boss.h**:   - Thêm lớp `Boss` thay thế cho `RoundKing`.   - Thêm thuộc tính tĩnh `bossesKilled` để theo dõi số lượng boss bị tiêu diệt.   - Cập nhật logic liên quan đến animation, máu, và sát thương của boss.
- **bullet.cpp / bullet.h**:
  - Cải tiến lớp `Bullet` để tương thích với lớp `Boss`.
  - Sửa đổi logic xử lý va chạm và sát thương để đảm bảo hoạt động chính xác.
  - Tối ưu hóa việc cập nhật animation và hộp va chạm (collision box).

- **CameraController.cpp / CameraController.h**:
  - Cải thiện khả năng theo dõi camera khi chiến đấu với boss.
  - Sửa lỗi nhỏ liên quan đến việc di chuyển camera và giới hạn biên.

- **hpBar.cpp / hpBar.h**:
  - Cập nhật lớp `HPBar` để hỗ trợ hiển thị thanh máu động cho boss.
  - Thêm khả năng tùy chỉnh màu sắc thanh máu cho nhiều boss cùng lúc.
  - Sửa lỗi quản lý bộ nhớ khi xóa thanh máu.

- **Tổng quan**:
  - Kiểm tra tính tương thích với các cơ chế hiện tại và đảm bảo không có lỗi phát sinh.

+++ Hack 'n' Slash/hpBar.cpp
@@ -1,50 +1,52 @@
 ﻿#include "hpBar.h"
 #include <algorithm>
 
+// Constructor: Khởi tạo thanh HP với các tham số ban đầu
 HPBar::HPBar(int initialX, int initialY, int width, int height, LivingEntity* target, SDL_Color barColor)
 {
-    x = initialX; // Tọa độ x ban đầu của thanh HP
-    y = initialY; // Tọa độ y ban đầu của thanh HP
+    x = initialX; // Tọa độ x ban đầu
+    y = initialY; // Tọa độ y ban đầu
     barWidth = width; // Chiều rộng của thanh HP
     barHeight = height; // Chiều cao của thanh HP
-    entity = target; // Con trỏ đến đối tượng LivingEntity mà thanh HP này theo dõi
-    color = barColor; // Màu của thanh HP (do người dùng chỉ định)
-    currentHp = target ? target->hp : 0; // Khởi tạo HP hiện tại: nếu target tồn tại thì lấy HP của target, không thì đặt bằng 0
+    entity = target; // Đối tượng LivingEntity mà thanh HP theo dõi
+    color = barColor; // Màu của thanh HP
+    currentHp = target ? target->hp : 0; // HP hiện tại, nếu không có đối tượng thì mặc định là 0
 }
 
+// Hàm vẽ thanh HP lên màn hình
 void HPBar::draw()
 {
-    if (entity != NULL) // Kiểm tra xem đối tượng được theo dõi có tồn tại không
+    if (entity != NULL) // Kiểm tra nếu đối tượng LivingEntity tồn tại
     {
-        // Giảm mượt mà currentHp để khớp với HP thực tế của entity
+        // Giảm dần currentHp để khớp với HP thực tế của đối tượng (tạo hiệu ứng mượt)
         if (currentHp > entity->hp)
         {
-            currentHp -= 0.5f; // Giảm dần currentHp (0.5f là tốc độ giảm, có thể điều chỉnh để mượt hơn)
-            if (currentHp < entity->hp)
-                currentHp = entity->hp; // Đảm bảo currentHp không nhỏ hơn HP thực tế của entity
+            currentHp -= 0.5f; // Giảm giá trị currentHp
+            if (currentHp < entity->hp) // Đảm bảo không giảm quá HP thực tế
+                currentHp = entity->hp;
         }
 
         // Vẽ viền ngoài của thanh HP
-        SDL_SetRenderDrawColor(Globals::renderer, 255, 255, 255, SDL_ALPHA_OPAQUE); // Đặt màu trắng cho viền
-        SDL_Rect barContainer = { x, y, barWidth, barHeight }; // Hình chữ nhật đại diện cho toàn bộ khung thanh HP
-        SDL_RenderDrawRect(Globals::renderer, &barContainer); // Vẽ viền khung
+        SDL_SetRenderDrawColor(Globals::renderer, 255, 255, 255, SDL_ALPHA_OPAQUE); // Màu trắng
+        SDL_Rect barContainer = { x, y, barWidth, barHeight }; // Hộp chữ nhật viền ngoài
+        SDL_RenderDrawRect(Globals::renderer, &barContainer); // Vẽ viền
 
-        // Tính tỷ lệ phần trăm HP còn lại của entity, giới hạn trong khoảng [0, 1]
-        float hpPercent = currentHp / (entity->hpMax * 1.0f); // Chia cho HP tối đa để ra phần trăm
-        hpPercent = std::max(0.0f, std::min(1.0f, hpPercent)); // Đảm bảo giá trị nằm trong [0, 1]
+        // Tính phần trăm HP còn lại
+        float hpPercent = currentHp / (entity->hpMax * 1.0f); // Tỷ lệ HP hiện tại so với HP tối đa
+        hpPercent = std::max(0.0f, std::min(1.0f, hpPercent)); // Đảm bảo giá trị nằm trong khoảng [0, 1]
 
-        // Thay đổi màu thanh HP dựa trên mức HP còn lại
-        if (hpPercent < 0.2f) // Nếu HP dưới 20%
+        // Thay đổi màu thanh HP dựa trên phần trăm HP còn lại
+        if (hpPercent < 0.2f) // Nếu HP còn dưới 20%
         {
-            SDL_SetRenderDrawColor(Globals::renderer, 255, 0, 0, SDL_ALPHA_OPAQUE); // Đặt màu đỏ để cảnh báo
+            SDL_SetRenderDrawColor(Globals::renderer, 255, 0, 0, SDL_ALPHA_OPAQUE); // Màu đỏ
         }
-        else
+        else // Nếu HP còn trên 20%
         {
-            SDL_SetRenderDrawColor(Globals::renderer, color.r, color.g, color.b, SDL_ALPHA_OPAQUE); // Dùng màu gốc của thanh HP
+            SDL_SetRenderDrawColor(Globals::renderer, color.r, color.g, color.b, SDL_ALPHA_OPAQUE); // Màu mặc định
         }
 
-        // Vẽ phần HP còn lại bên trong khung
-        SDL_Rect hpRect = { x + 2, y + 2, (barWidth - 4) * hpPercent, barHeight - 4 }; // Hình chữ nhật biểu thị HP còn lại, thu nhỏ 2 pixel mỗi bên để vừa viền
-        SDL_RenderFillRect(Globals::renderer, &hpRect); // Tô màu cho phần HP còn lại
+        // Vẽ phần thanh HP đã được lấp đầy
+        SDL_Rect hpRect = { x + 2, y + 2, static_cast<int>((barWidth - 4) * hpPercent), barHeight - 4 }; // Hộp chữ nhật bên trong
+        SDL_RenderFillRect(Globals::renderer, &hpRect); // Vẽ phần lấp đầy
     }
 }
